HGSS Fishsanity: Add Routesanity
Adds Routesanity as an option for HeartGold & SoulSilver Fishsanity.
Enabling Routesanity also adds a variety of different items to the pool, including a new Progression item in the SecretPotion, required to access the Cliff Edge Gate. Changes have been designed so as to not impact Routesanity being disabled.
This also adds 'event' locations for catching each Pokemon for the first time, whether or not Routesanity has been enabled. The victory condition has been changed to requiring all of the 'events'.