KTD: Precollect Stage 1
Adds the randomized Stage 1 to the player's starting inventory instead of putting it on a location.
This was done in the hopes that it'll make the randomized stage mechanic a little clearer to new players, as well as in the hope that the slightly opened up sphere 1 will make it a little more fun to play.
Stage 1 in particular is precollected instead of the entire set of Level 1 Stages due to it helping to maintain an 'intended order' to Stages instead of letting the player choose which order to do them in, which is closer to the original game.