Fully implemented an option for adjusting excess Sun Stones
Added an option called "Excess Sun Stones" (`excess_sun_stones`, ExtraSunStones) that can be used to modify the classification or existence of Sun Stones that aren't required to reach the player's goal.
Available options are "Progression" (default behavior, leaves all Sun Stones as progression items), "Useful" (changes the excess to the classification of its namesake), "Filler" (same as 'Useful') and "Removed" (completely removes the excess Sun Stones from the item pool).
If no Sun Stones are required for any boss, they're automatically classified as "Filler", ignoring if the player wanted them to be "Progression" or "Useful" (because they would serve literally no purpose...). Might change this to only change "Progression" to "Filler" in the future. Also, "Removed" takes priority in this case, so the excess will be removed in addition to the classification being changed (to account for item links and plando).

Finally, renamed `enable_kirby_fighters_locations` to just `kirby_fighters_locations`. An error is now thrown if the old option name is enabled; might change this in the future to check for if it exists at all.