KTD: Minor changes, option description cleanup, and option presets
Removed the "Unlock Copy Ability Testing Room" location's category so that it should now be displayed alongside the victory location.
I intend to change this back if at some point in the future the ability to put the victory location in a category is added, but the intention is for them to always be together.

Changed Grand Sun Stones' classification to Progression + Useful.
Adjusted most option descriptions to better suit WebHost.

Moved the `excess_sun_stones` option further down, both due to it being an advanced option and because it didn't display properly on WebHost in its prior position.