KTD: Update LL1 Sun Stone 2 normal logic
Puts Parasol in logic for the second Sun Stone in Lollipop Land Stage 1 when using normal logic due to the discovery of a consistent method for it.

Wheel is still clearly the intended method so easy logic is unchanged, and Parasol was already in logic on hard since hard logic only needs the player to have any ability.