KTD: Improve logging, add category to ATR location
Switches the prior `print` statements over to `logging.warning` and adds extra details, such as the relevant player and more specific explanations of the changes.

Also moves the Unlock Copy Ability Testing Room location to its own category so that it'll be next to the victory button in the client.