Updated documentation
Changed the bug report link to go to the issues page instead of the Manual Discord, in an attempt to prevent people from pestering the people there from being bothered by someone who might have been directed there and isn't aware of the thread system. It probably won't ever happen, but on the off-chance that it does... (Tech support still links to the Discord, since I can actually mention the thread there.)

Added a list of the items that can be randomized to the game info page, for greater clarity.

Adjusted a couple of sentences.