Update to Manual Unstable 20251205
Updated to Manual Unstable 20251205.
Refactored some hooks to now properly use the newly-added `before_generate_early`.
Added the new `sort-key` tag to many locations in `locations.json` in order to maintain the original spirit of the order.
Added a new `lose_ability_traps` option in place of the old `filler_traps`. It does the same thing but with a different name so that I can more freely customize it. Backwards compatibility with `filler_traps` was retained.